created, EffectAccumulator, moved appropriate effect-related files to their own package
